+++ src/lib/auth.js
@@ -1,4 +1,4 @@
-import client from "../client";
+import client from "./client";
 
 
 export const register = (credentials) => {
