+++ src/lib/room.js
@@ -1,4 +1,4 @@
-import client from "../client";
+import client from "./client";
 
 export const fetchRooms = () =>
     client.get(`/rooms`);
